Personal Site/requirements.txt
flask
gunicorn
cryptography
flask-socketio
eventlet==0.30.2
Flask-Failsafe
Flask-Cors
mysql-connector-python
